Moved mprans2/setup.py into proteus/setup.py
