--- proteus/mprans2/setup.py
@@ -1,164 +1,0 @@
-from setuptools import setup, Extension
-from setuptools.command.build_ext import build_ext
-import sys
-import os
-import setuptools
-
-__version__ = '0.0.1'
-
-
-class get_pybind_include(object):
-    """Helper class to determine the pybind11 include path
-
-    The purpose of this class is to postpone importing pybind11
-    until it is actually installed, so that the ``get_include()``
-    method can be invoked. """
-
-    def __init__(self, user=False):
-        self.user = user
-
-    def __str__(self):
-        import pybind11
-        return pybind11.get_include(self.user)
-
-
-class get_numpy_include(object):
-    """Helper class to determine the numpy include path
-
-    The purpose of this class is to postpone importing numpy
-    until it is actually installed, so that the ``get_include()``
-    method can be invoked. """
-
-    def __init__(self):
-        pass
-
-    def __str__(self):
-        import numpy as np
-        return np.get_include()
-
-
-ext_modules = [
-    Extension(
-        'cAddedMass',
-        ['AddedMass.cpp'],
-        depends=['AddedMass.hpp'],
-        include_dirs=[
-            # Path to pybind11 headers
-            get_pybind_include(),
-            get_pybind_include(user=True),
-            get_numpy_include(),
-            os.path.join(sys.prefix, 'include'),
-            os.path.join(sys.prefix, 'Library', 'include'),
-            '../'
-        ],
-        language='c++'
-    ),
-    Extension(
-        'SedClosure',
-        ['SedClosure.cpp'],
-        depends=['SedClosure.hpp'],
-        include_dirs=[
-            # Path to pybind11 headers
-            get_pybind_include(),
-            get_pybind_include(user=True),
-            get_numpy_include(),
-            os.path.join(sys.prefix, 'include'),
-            os.path.join(sys.prefix, 'Library', 'include'),
-            '../'
-        ],
-        language='c++'
-    ),
-    Extension(
-        'cVOF3P',
-        ['VOF3P.cpp'],
-        depends=['VOF3P.hpp'],
-        include_dirs=[
-            # Path to pybind11 headers
-            get_pybind_include(),
-            get_pybind_include(user=True),
-            get_numpy_include(),
-            os.path.join(sys.prefix, 'include'),
-            os.path.join(sys.prefix, 'Library', 'include'),
-            '../'
-        ],
-        language='c++'
-    ),
-    Extension(
-        'cNCLS3P',
-        ['NCLS3P.cpp'],
-        depends=['NCLS3P.hpp'],
-        include_dirs=[
-            # Path to pybind11 headers
-            get_pybind_include(),
-            get_pybind_include(user=True),
-            get_numpy_include(),
-            os.path.join(sys.prefix, 'include'),
-            os.path.join(sys.prefix, 'Library', 'include'),
-            '../'
-        ],
-        language='c++'
-    ),
-]
-
-
-def has_flag(compiler, flagname):
-    """Return a boolean indicating whether a flag name is supported on
-    the specified compiler.
-    """
-    import tempfile
-    with tempfile.NamedTemporaryFile('w', suffix='.cpp') as f:
-        f.write('int main (int argc, char **argv) { return 0; }')
-        try:
-            compiler.compile([f.name], extra_postargs=[flagname])
-        except setuptools.distutils.errors.CompileError:
-            return False
-    return True
-
-
-def cpp_flag(compiler):
-    """Return the -std=c++14 compiler flag  and errors when the flag is
-    no available.
-    """
-    if has_flag(compiler, '-std=c++14'):
-        return '-std=c++14'
-    else:
-        raise RuntimeError('C++14 support is required by xtensor!')
-
-
-class BuildExt(build_ext):
-    """A custom build extension for adding compiler-specific options."""
-    c_opts = {
-        'msvc': ['/EHsc'],
-        'unix': [],
-    }
-
-    if sys.platform == 'darwin':
-        c_opts['unix'] += ['-stdlib=libc++', '-mmacosx-version-min=10.7']
-
-    def build_extensions(self):
-        ct = self.compiler.compiler_type
-        opts = self.c_opts.get(ct, [])
-        if ct == 'unix':
-            opts.append('-DVERSION_INFO="%s"' % self.distribution.get_version())
-            opts.append(cpp_flag(self.compiler))
-            if has_flag(self.compiler, '-fvisibility=hidden'):
-                opts.append('-fvisibility=hidden')
-        elif ct == 'msvc':
-            opts.append('/DVERSION_INFO=\\"%s\\"' % self.distribution.get_version())
-        for ext in self.extensions:
-            ext.extra_compile_args = opts
-        build_ext.build_extensions(self)
-
-setup(
-    name='mprans2',
-    version=__version__,
-    author='JohanMabille',
-    author_email='johan.mabille@gmail.com',
-    url='',
-    description='',
-    long_description='',
-    ext_modules=ext_modules,
-    install_requires=['pybind11>=2.0.1', 'numpy'],
-    cmdclass={'build_ext': BuildExt},
-    zip_safe=False,
-)
